requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
openai==1.3.5
langchain==0.0.340
langchain-community==0.0.10
langchain-openai==0.0.2
pypdf2==3.0.1
python-docx==1.1.0
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
